refactor(coordinator): extract API key handlers from consumer.go
Move the consumer-facing API key management endpoints and their
validation/response helpers (create/list/get/update/rotate/delete,
apiKeyToResponse, validateKeyLimitInputs, keyModelAllowed, checkKeySpendCap,
applyKeyPatch) into apikey_handlers.go, matching the existing
apikey_handlers_test.go. Same package, no behavior change. consumer.go drops
from ~5311 to ~4884 LOC.
